License update.
Changed from New BSD to a custom license. The new license requires a Spine license to use the code. If you have a valid Spine license, you can do whatever you like with the code. This should not be a problem for anyone using the runtimes with Spine, nothing changes. If using the runtimes without a Spine license, you now need a Spine license. This is because the runtimes were created explicitly to be used with Spine.